automations/invoice-webapp/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pdfplumber==0.10.3
python-dotenv==1.0.0
requests==2.31.0
